[FAB-1360] Introduce ChainPartition for Kafka
https://jira.hyperledger.org/browse/FAB-1360

This changeset introduces ChainPartition, a construct that will be used
to identify the Kafka topic/partition that the ordering shims should
interact with when dealing with a particular chain.

Note that it encodes the given chain ID in base-16 as a means of escaping
characters that are not allowed in Kafka topic names. All the
restrictions for topic names are linked to in line 38. Filtering for
complying chain IDs should probably happen during transaction filtering
at ingress. This is marked as a TODO, and a JIRA issue has been created
for it.

This changeset merely introduces the ChainPartition construct and does
not integrate it with the rest of the code yet. This will happen in a
follow-up changeset.

Change-Id: I33d4e80758d7dfd7e7d827caa385d0507a39a40a
Signed-off-by: Kostas Christidis <kostas@christidis.io>