Refactor MSP package and msp config w/o json
This change-set presents a better organisation of the MSP package: code that
belongs to msp proper is in fabric/msp, whereas core that instantiates and
manages msps for the peer was moved in fabric/core/peer (a new package was
created to avoid import cycles). Furthermore, configuration of the local
MSP no longer requires a specific json file: it is sufficient to place
certificates and keys in a directory with the appropriate structure (sample in
msp/sampleconfig).

Change-Id: Ic0c696a6e0fb406d8e482240e24e5d5774efc5c5
Signed-off-by: Alessandro Sorniotti <ale.linux@sopit.net>