[FAB-1363] Move ChainID method to ConsenterSupport
https://jira.hyperledger.org/browse/FAB-1363

This changeset moves the ChainID() method from the ChainSupport interface
to the ConsenterSupport one. What necessitates this move is that the
Kafka multichain.Chain object returned/expected by HandleChain(), needs
to be able to allocate the resources necessary to keep up with its
respective chain when the Start() method is invoked by the multichain
manager, as the contract of the Chain interface (defined in
chainsupport.go) dictates.

Keeping up with the respective chain means that the Kafka
multichain.Chain object needs to bring up a producer/consumer that
connects to the corresponding chain partition, so it needs to know the
chain ID. The only object passed to us during the HandleChain call is an
object of type multichain.ConsenterSupport. Therefore, this interface
needs to be extended with the ChainID() method.

I've noted the need for this during my review of this changeset:
https://gerrit.hyperledger.org/r/#/c/2763/ (skip to the comment
timestamped 11-27 18:52, and its follow-up comment)

If you want to see how this will look in action for the Kafka consenter
(coming in a follow-up changeset), please see:
https://github.com/kchristidis/fabric/blob/47752ed61fcab1b26207a9e9075c1c793d723912/orderer/kafka/main.go#L128
https://github.com/kchristidis/fabric/blob/47752ed61fcab1b26207a9e9075c1c793d723912/orderer/kafka/main.go#L143

Change-Id: I80590534097a1f4b0f30267ee541a3d3e7f7e3e1
Signed-off-by: Kostas Christidis <kostas@christidis.io>