FAB-1378 beginnings of a join command
https://jira.hyperledger.org/browse/FAB-1378

"peer node start" works as before and creates the default chain.

"peer node start --peer-defaultchain=false" however starts without
a chain. It just creates the gossip service and starts up the
peer server in a chainless, ledgerless mode. The peer is read to
accept join calls.

"peer node join -b <path to genesis block file>" calls CSCC with
the genesis block.

kvledgers is removed and the peer package is used for ledger access.

CSCC is under construction but this CR allows the work to be driven
via CLI (and later from SDK).

Change-Id: Id581c1f04b8788f54be467f593d74ea6b1efe713
Signed-off-by: Srinivasan Muralidharan <muralisr@us.ibm.com>