MSP mgr instantiation from Block
This change-set introduces a proper implementation of the function that
generates an MSP manager out of a Block from the ledger. We expect one
ConfigurationItem per MSP, each containing in the Value field the protobuf
serialization of an MSPConfig message. The MSP config schema has been further
simplified by removing the MSPManagerConfig struct. Furthermore, MSP's and
MSPManager's are now immutable types: the reconfig method has been removed and
reconfiguring a manager/msp means constructing a new one and letting the old
instance go out of scope.

Change-Id: I42b12e1bece4084ab5695f60050ce512a5286e87
Signed-off-by: Alessandro Sorniotti <ale.linux@sopit.net>