[FAB-1410] Remove node's call to bootstrapper
https://jira.hyperledger.org/browse/FAB-1410

The node package gets the genesis block by invoking the static
bootstrapper, then uses that block for this call:

    `deliverService.JoinChannel(commit, block)`

It should be noted that this is a temporary solution by the node package
authors to test out the end-to-end integration.

However, this layering violation (which could be ignored temporarily due
to the reason cited above) manifests itself in a very practical and real
way:

1. The static bootstrapper is being replaced by the provisional
bootstrapper, for the reasons documented here:
https://jira.hyperledger.org/browse/FAB-1364.
2. The provisional bootstrapper loads the `orderer.yaml` file (in the
`orderer` package) to determine its output.
3. The fabric-peer Docker image however does not have access to this file.
4. Thus, if we call the provisional bootstrapper from the node, the
`fabric-peer` container will crash.

Upon further inspection, it seems that what the node really needs now is
a configuration block with just the right chain ID, on the payload header
of its first enclosed envelope.

The simplest, least invasive way to offer this is by creating a sample
block via the following call:

    `utils.MakeConfigurationBlock(util.GetTestChainID())`

Change-Id: Ifb7e5b1617c0b0d49eff2598dfdb1bfd0628bf1f
Signed-off-by: Kostas Christidis <kostas@christidis.io>