PKCS11/MSH compatible BCCSP SKI gen
This change-set introduces a new way to compute the SKI
of bccsp keys that is based on what happen in the PKCS11/HSM
world. The SKI of a private key is compute as SHA-256
of the der representation of the corresponding public key.
For AES keys, they are prepended by 0x01, for scoping,
and then hashed.

Change-Id: Id4c5d4a80bdfe07feb39bebd235e0625994bf91c
Signed-off-by: Angelo De Caro <adc@zurich.ibm.com>