[FAB-1419] Set appropriate ACK level for brokers
https://jira.hyperledger.org/browse/FAB-1419

By default the partition leader (i.e. the Kafka broker responsible for
maintainig a chain) waits only for the local commit to succeed before
responding.

The right setting to maximize the reliability of data delivery is to have
the partition leader wait for all in-sync replicas (backups) to get a
message before sending back to the producer an acknowledgement or an
error.

Change-Id: I1bc045ef883c39cac3fec2c31f3e9f329ba972c4
Signed-off-by: Kostas Christidis <kostas@christidis.io>