[FAB-1351] New chain config client for Kafka
https://jira.hyperledger.org/browse/FAB-1351

As things stand right now, the configuration transaction posted by the
broadcast_config sample client only works for solo. This is because the
genesis block for the Kafka orderer contains extra keys.

This changeset:

1. Introduces an extra flag that allows us to specify the consenter
type, and then, by using the provisional bootstrapper, it creates the
appropriate configuration transaction.

2. Refactors/simplifies the client to bring it more in line with the rest
of the sample_clients in the `orderer` package and make future
maintenance easier.

Change-Id: I61ce5dfb36a11ab58d41a65267768cf0948263dc
Signed-off-by: Kostas Christidis <kostas@christidis.io>