[FAB-1420] Convert signature policy to provider
https://jira.hyperledger.org/browse/FAB-1420

The cuathdsl signature policy type used to be the only policy type
available.  In order to allow new types of policy, a provider framework
is needed.

This changeset introduces the provider framework and converts the
existing cauthdsl based signature policy work to be one of these
providers.

Finally, in response to some changeset feedback, it also introduces the
notion of a SignedData struct which can be common to all of the crypto
implementations (and signficiantly simplifies the interface
definitions).

Change-Id: Ib67ff9f6de0433f73c96e5a4b229f1587a0f0363
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>