Chain MSPManager and chain ACLs config schema
This is a changeset suggesting MSPManager & chain ACL related configuration
schema for a new chain/channel. This is put under protos/common,
as it is a configuration relevant to orderer channel setup
and application channels.

This changesets suggests a representation for chain readers
(identities that are allowed to read a chain), chain writers (identities
that are authorized to submit transactions to a chain), and chain admins
(identities that are authorized to administer a chain). The structures
defined here can be used in conjuction to the policy framework.

Files in this changeset:
 - common/chain-config.proto: definition of a config schema for chain
   MSPs, as well as the chain readers, writers and admins
 - common/chain-config.pb.go: its golang version

Change-Id: Ia8fe10e6d0a3db5e24a502cc5edcbe59cd6ca920
Signed-off-by: Elli Androulaki <lli@zurich.ibm.com>