[FAB-1433] (BUGFIX) Add batch timeout in genesis
https://jira.hyperledger.org/browse/FAB-1443
https://gerrit.hyperledger.org/r/#/c/3461/

The above changeset fixed a TODO to have the batch timeout retrieved via
the shared config mechanism from the genesis block, but it did not
actually encode the batch timeout into the genesis block.  This was
causing a batch timeout of 0s and was breaking the bdd tests.

This changeset adds the encoding of the batch timeout to the provisional
bootstrapper.

Change-Id: I938b4aa89b2a2a5d3ca049129a32b3a8ec37832e
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>