Nominate Jim Zhang to be Fabric Maintainer
Jim has been leading the fabric-sdk-node development and has
been conducting a number of consistently critical reviews
for that and other fabric projects. He led the development
of the Fabric SDK specification and is actively working with
the python and java SDK teams.

I think he'd make an excellent addition to the team.

Change-Id: Ibb403adcf66fb3cdd014e3178cfb9ff3ecff2238
Signed-off-by: Christopher Ferris <chrisfer@us.ibm.com>