[FAB-1521] Fix rawledger to support restart
https://jira.hyperledger.org/browse/FAB-1521

The rawledger interface was originally implemented as a single chain
initialized with a genesis block. This causes problems on restart
because a newer genesis block may already exist and passing the genesis
block becomes nonsensical.

This changeset changes this behavior to intialize no ledgers on factory
creation, and to allow the use of Append for blocks, rather than
transaction sets.  This satisfies a secondary goal of allowing block
creation without block commit, as is required by the SBFT flow of
consenting on blocks, rather than on transactions.

Change-Id: Icbb0bb1a80f9e04861f4302966709f1752068961
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>