[FAB-1523] Populate block metadata LastConfig
https://jira.hyperledger.org/browse/FAB-1523

This changeset adds new data types for Metadata and for
LastConfiguration which is the second field of the block metadata.

The Metadata type is intended to accomodate the other metadata fields as
well, but is currently only utilized for the LastConfiguration.

This changeset also populates the LastConfiguration field of the block
metadata with this new data structure.

Change-Id: I36e72f4c27b67dd7455aae2f423b4b14e54f9413
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>