Removing core/chaincode/shim/crypto
This change-set removes the crypto package under
core/chaincode/shim. All of its functions have
been moved to the accesscontrol package and
the chaincode examples using it have been updated
to use the new package.

Notice that the accesscontrol package is still
under development. More change-sets will be
submitted to shape it.

This change-set comes in the context of
https://jira.hyperledger.org/browse/FAB-831

Change-Id: I42bc59265f1a84aae71e971c063d119a410eb361
Signed-off-by: Angelo De Caro <adc@zurich.ibm.com>