[FAB-1617] [FAB-1619] Utilize configtx.Template
https://jira.hyperledger.org/browse/FAB-1617
https://jira.hyperledger.org/browse/FAB-1619

This changeset removes some of the duplicated orderer bootstrap code
from the protos/utils package.  It does this by introducing a
common/configtx/test package which uses a sample config template to
produce a genesis block.

This can also be used in the future for clients wishing to create a
chain creation transaction.  This is demonstrated in the
broadcast_config client.

This removes the vestigal utils methods for creating configuration
transactions via the 'oldConfiguration' mechanisms.

Change-Id: If064ed3cfcdbd2e59e8fb6c7bf72c8af160dc5f6
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>