[FAB-1659] Update BDD Compose ENV vars
https://jira.hyperledger.org/browse/FAB-1659

In FAB-1612 [1] new configuration keys were introduced on the
`orderer.yaml` file. The ENV vars on the Compose files we use for the
orderer-related BDD tests however were not updated accordingly. As a
result, all tests would run against the default config setting
(read: solo) and the Kafka path would not get tested.

This changeset addresses this issue.

Additionally it:

1. Removes all environment variables from the Compose files when these
match the defaults in `orderer.yaml`. This reverses an earlier take on
the matter. The thinking behind this change is that we don't need to
keep two sets of settings unless absolutely necessary.

2. Formats `orderer.yaml` so that the recently-added lines break at 80
characters, as is the case with the rest of that file.

[1] https://gerrit.hyperledger.org/r/#/c/3879/

Change-Id: I0d1e137bd4a8e523edb7cb6a060c09ff80d31340
Signed-off-by: Kostas Christidis <kostas@christidis.io>