[FAB-1662] Rename main kafka package file
https://jira.hyperledger.org/browse/FAB-1662

In the `kafka` package, most of the logic that ties everything together
resides in `main.go`. This changeset renames this file to `orderer.go`,
to avoid any connotations with `main` which in Go pretty much implies a
binary. (Bonus points for the new name being a bit more descriptive as
well.)

Change-Id: Ie9b55f7fa081e20cf8affb1cdc8bd1315a216de6
Signed-off-by: Kostas Christidis <kostas@christidis.io>