FAB-1505 and FAB-1337 Refactor CouchDB code
FAB-1505 refactors CouchDB code to be aligned with LevelDB structure.
FAB-1337 applies all ledger tests written for LevelDB against CouchDB.

In order to test the CouchDB refactor,
had to implement FAB-1337 in same changeset to test.

Like most refactors, this changeset touches a lot of files and code,
since it is not feasible to do a partial refactor.

Tests related to versioning and deleting have been commented out,
these will be re-enabled in subsequent changeset when CouchDB
has the support, in order to keep this changeset size more reasonable.

Change-Id: I0d2d6cca89bd0252f6e84317457a9140b3a0d7a5
Signed-off-by: denyeart <enyeart@us.ibm.com>