[FAB-1639] [FAB-1580] Rework validator
This change-set has removed calls to the txvalidator that were issued right
after the peer (the leader) receives blocks from the orderers. The validator
is now called to validate messages received from the gossip layer. In order
to fix an import cycle, we have introduced the ChaincodeProvider interface
in core/common/ccprovider/ccprovider.go, an implementation and a factory.
This way, code that needs to use functions from the chaincode package
without importing it can simply use (and possibly extend) the
ChaincodeProvider interface and implementation.

Furthermore, this drop has introduced protocol-level message validation for
config transactions that was lacking before.

Change-Id: I5906a6fe3da8410b05b5079dd7f0b9d28d20bb85
Signed-off-by: Alessandro Sorniotti <ale.linux@sopit.net>