FAB-1665 go chaincode does not need core.yaml
https://jira.hyperledger.org/browse/FAB-1665

All parameters are passed to chaincode via environment
variables or as parameters. Default properties are not
picked up from core.yaml. In 1.0 architecture, the peer
no longer builds the chaincode package but only passes it
to docker to deploy (the package is built by CLI or SDK).
In this mode the core.yaml in the Dockerfile will typically
not be available in the peer environment.

Change-Id: Ifffece4edb4fda103c379b644327972429034a2b
Signed-off-by: Srinivasan Muralidharan <muralisr@us.ibm.com>