[FAB-187] - using policies in VSCC
This code drop integrates cauthdsl into the VSCC. The flow is the following:
the validator code extracts from LCCC the policy specified by the deployer.
The policy is an additional argument to VSCC. VSCC deserializes the policy
and uses cauthdsl to validate the endorsements against it.

Currently the chaincode deployer doesn't specify a policy when it deploys a
chaincode and so for now the validator uses a policy that says "signed by any
member of the specified MSP". This is to be removed as soon as chaincode
deploy transactions specify policies when a new chaincode is deployed.

Change-Id: I9e5ff1effa75934cdfef231a37e66f2941a45d10
Signed-off-by: Alessandro Sorniotti <ale.linux@sopit.net>