orderer.template needs to be regenerated again
Though this is a temporary solution for playing with create/join
chain, its important to keep this up-to-date.

For future reference, when the config template mechanism is changed
we need to make sure to regenerate orderer template using tooling in
orderer/tools/configtemplate/ and copy orderer.template to
common/configtx/test/orderer.template.

Change-Id: I82e07eed9bb83a1a11766d6a1dea70a76097aa4e
Signed-off-by: Srinivasan Muralidharan <muralisr@us.ibm.com>