[FAB-1697] Add chain hashing algorithm config
https://jira.hyperledger.org/browse/FAB-1697

This changeset adds the chain hashing algorithm to the configuration
proto.  It also adds a chain configuration consumer (for configuration
items of chain type) which is capable of processing this message once
embedded in the genesis block.

Actually embedding this and utilizing the chain configuration will come
in later changesets.

Change-Id: I960ce4c6ca73b988840585364876e970cbf99a47
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>