[FAB-1747] Do not stop devmode chaincode after deploy
The normal flow for chaincode deployment involves an
implicit stop of the container as subphases of deployment.
However, devmode chaincode is primarily under the control
of a user, not the peer.  Therefore, the current peer
is broken for devmode because it tries to shutdown
the container after registration.  Because the chaincode
was started by the user, the stop ultimately fails but not
before the peer has confused its internal state by marking
it as down.

The primary problem is that subsequent calls into the
chaincode will ultimately fail on account of the quasi
running-but-not-really-running state and an inability
to try to correct the problem autonomously due to the
lack of control over the users process.

This patch addresses all of the above by acnowledging that
devmode chaincode should simply be left in a running
state.

Change-Id: I2e8f1f81a33019a10e3e65d1416ea4426a19e77a
Signed-off-by: Greg Haskins <gregory.haskins@gmail.com>