FAB-1763 Rename **TEST_CHAINID** to test_chainid
Chainid becomes database name in CouchDB.

CouchDB naming conventions:

Only lowercase characters (a-z), digits (0-9), and any of the
characters _, $, (, ), +, -, and / are allowed. Must begin
with a letter.

Therefore, we either need to adopt these same restrictions for
chainid naming, or have a deterministic mapping of chainid to
a valid (and unique) CouchDB database name.

The complete fix will be delivered in FAB-1764.  In FAB-1763
we simply change '**TEST_CHAINID**' to 'test_chainid' in order
to unblock current failures.

Change-Id: Id274b37f5a07813b172db0e695ec803e63a09726
Signed-off-by: denyeart <enyeart@us.ibm.com>