[FAB-1773] Add metadata field for orderer use
https://jira.hyperledger.org/browse/FAB-1773

A consensus implementation may need to read the ledger and extract
metadata from it during operation (most likely when booting up).

As an example, the Kafka-based orderer, should read the offset of the
last envelope it placed into a block *and* wrote to the local ledger, and
use that value to resume consumption of chain's partition.

This changeset:
1. Introduces a metadata field for the described use above
2. Sizes and populates the `Metadata` field of a newly-constructed block
appropriately.

These changes will be put into use in a follow-up changeset [FAB-1774].

Change-Id: Ib3d7d8f77c1a888ebec5790a144457913b201202
Signed-off-by: Kostas Christidis <kostas@christidis.io>