[FAB-1623] Add restart support to Kafka orderer
https://jira.hyperledger.org/browse/FAB-1623

Allow a restarted ordering shim to pick up ordering exactly from where
it left it off.

This changeset:

1. Adds a `KafkaMetadata` proto message definition to hold the
orderer-related metadata that will be written with every block on the
local ledger.
2. Modifies the Kafka-based orderer so that (a) it reads that info upon
booting up and (b) it then seeks to the appropriate offset.

A convoluted and somewhat tautological unit ("unit") test has been added
to test this functionality, but we are really starting to hit the limit
of what we can do with the provided mock structures. This path will be
revisited with BDD tests within the next few weeks.

I've also renamed the offset-related variables in all of the tests to
make things easier to keep track of.

Review starting point:
- fabric/protos/orderer/kafka.proto
- fabric/orderer/kafka/orderer.go

[1] https://jira.hyperledger.org/browse/FAB-1773
[2] https://jira.hyperledger.org/browse/FAB-1623

Change-Id: I1314a2361844dd4adbd78bf57bb08aaee358dc6f
Signed-off-by: Kostas Christidis <kostas@christidis.io>