[FAB-1774] Use metadata field for orderer info
https://jira.hyperledger.org/browse/FAB-1774

A consensus implementation may need to read the ledger and extract
metadata from it during operation (most likely when booting up).

As an example, the Kafka-based orderer, should read the offset of the
last envelope it placed into a block and wrote to the local ledger, and
should use that offset to resume consumption of chain's partition.

This changeset follows up on the work of FAB-1773 [1].

Specifically, it modifies:

1. The common components `Consenter` interface so that the
newly-introduced metadata field is passed on to the consensus
implementations via the `HandleChain` method.
2. The `WriteBlock` method of the `ConsenterSupport` interface so that
this metadata can be persisted to blocks.

It also adds relevant unit tests.

This is a precursor to FAB-1623 [2], which will add restart support to
the Kafka-based orderer.

Review starting point: fabric/orderer/multichain/chainsupport.go

[1] https://jira.hyperledger.org/browse/FAB-1773
[2] https://jira.hyperledger.org/browse/FAB-1623

Change-Id: I3d1c932eb30537f6a1aa8056b9d38550b17dee6d
Signed-off-by: Kostas Christidis <kostas@christidis.io>