[FAb-1699] Deterministic hashing for block header
https://jira.hyperledger.org/browse/FAB-1699

Currently, we compute the hash of the block header by hashing over the
protobuf marshaling of the BlockHeader object.  This is not guaranteed
to be deterministic, so a reliable marshaling scheme is required.

This changeset causes block headers to be marshaled to bytes via ASN.1.
Note: Due to an idiosyncracy in the golang ASN.1 implementation, block
numbers of MaxInt64 cannot be encoded, and this will result in a panic.

This resulting byte slice is used as the parameter to the hashing
algorithm.

Change-Id: I3b0582e89c17c0d120f3e2d888a2205022f2de18
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>