[FAB-1812] Pass configtx.Manager.Apply to validate
https://jira.hyperledger.org/browse/FAB-1812

The peer needs to apply configuration transactions as they are
committed.  This changeset hooks the configtx.Manager.Apply method into
core/committer/txvalidator/validator.go so that configuration is updated
as configuration blocks are committed.

Because it was in the same path, it also passes the MSPManager to the
validator to remove one of the dependencies on the throwaway MSP
multichain manager.

Change-Id: I4d606d850dddb684468d7a5dd7257e8dd967ad28
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>