[FAB-1444] Move signature to top level in gossip msg
A gossip.GossipMessage is a oneof of different types of messages,
and all messages in gossip are under that message.
Messages that are signed have their signature in the inner field of the message.
To make things more secure, and also more uniform, I'm moving the signature
to the top level of the GossipMessage.

Change-Id: Ia358015ed98270d882c3facdaca89355666a9f69
Signed-off-by: Yacov Manevich <yacovm@il.ibm.com>