Stop logging the gossip content message
Currently for debug purposes there is a log printout
which prints gossiping message from the delivery service
client. This is usefull while doing small debuging and
testing, while in heavy load it garbages the console.

This commit is to stop printing this message to reduce
log output and make it more usefull.

Change-Id: I20e3c38e9076cec09fd9438b42d3a56882399b54
Signed-off-by: Artem Barger <bartem@il.ibm.com>