[FAB-1790, FAB-1791] Chaincode calling chaincode
This changeset enables chaincode calling chaincode on the same channel
or different channel read-only. The called chaincode operates in the
same transaction context of the caller chaincode. See the chaincode
API for details.

Change shim.InvokeChaincode API to accept a channel.

Fix up the chaincode_example04 so that the current unit tests can
exercise both query and invoke.

Also update a number of files for better logging, including removing
repeating log in loop and info log in heavy transaction path.

Change-Id: I9b67b91311482d7c47cf66a0849c2f29594621f5
Signed-off-by: Binh Q. Nguyen <binhn@us.ibm.com>