FAB-1920 install a chaincode on local peer
https://jira.hyperledger.org/browse/FAB-1920

peer chaincode install -n <name> -p <path> -v <version>

This command should be executed on the local peer as opposed
to other commands such as "invoke", "query" which are sent to the remote peer.
ie, assumes the chaincode is on the peer (like chaincode_example02). Contrast
this with the old deploy that can be executed from any CLI/SDK outside of the
peer.

On successful execution the command will create a chaincode package on the file
system of the peer prefixed by the "fileSystemPath" as defined in core.yaml.

------- NOTE ------
-The "package" will eventually have more contents such as policy and will be
-signed etc, but for now is just the serialized chaincode deployment spec. This
will be used in subsequent CRs to get the mechanics of a chaincode on multiple
chains ironed out.
------- NOTE ------

Example
=======
With default "fileSystemPath: /var/hyperledger/production",

    peer chaincode install -n mycc -p my/chain/code -v firstversion

will create chaincode package in
    /var/hyperledger/production/chaincodes/mycc.firstversion
using the chaincode defined in my/chain/code.

Change-Id: I479834736d816ed522688dacae6e946340debf11
Signed-off-by: Srinivasan Muralidharan <muralisr@us.ibm.com>