Connect MSP-components-for gossip with gossip
The gossip layer uses 2 interfaces for identity
and cryptographic capabilities, which are implemented
using the various MSP managers (either local or channel-based).

The 2 interfaces are found under gossip/api
and they are called MessageCryptoService and SecurityAdvisor.

They were implemented using the MSP, and this commit
connects their implementation to the gossip layer,
and removes the fake/dummy implementations that were
used as placeholders.

- This test also fixes a bug that is related to parsing
the certificate of anchor peers.

- It also now makes the default anchor peer (if non specified) to be
peer0 with the certificate of msp/sampleconfig

- And it also fixes a small bug was introduced by a commit a few days ago
  that was caused by passing a certificate instead of a PKI-ID
  into the gossip identity mapper, when validating
  a channel state-info message.

How was this tested?
I ran 4 instances of peers, in 2 different orgs, but with rootCAs
of all orgs in all peer local MSP configurations, and with peer0
as a bootstrap peer for the rest 3 peers, and made sure that the gossip
communication logs in all 4 peers send messages to 3 other peers.
This proves that they are able to authenticate and exchange membership
information about each other.
Then I created a channel, and made peers join it, and saw that they send
one another channel related messages and no warnings/errors were seen
in the logs.

Signed-off-by: Yacov Manevich <yacovm@il.ibm.com>
Change-Id: Id8b5e7e9400b7c58302e4396c04209f917bbde70