Refactor identity code and check for MSP ID
This change set refactors the code inside of the MSP that is called whenever
a serialized identity has to be turned into an identity instance. The code
was previously repeating the deserialization into a SerializedIdentity struct
twice - once by the MSPManager and once by the MSP. Now the MSP
implementation exposes a function that turns SerializedIdentity.IdBytes
into an identity. This method is called by the public method that takes the
full serialized identity, deserializes it into a SerializedIdentity struct
and calls the private method passing in SerializedIdentity.IdBytes.

Also, the code now checks that SerializedIdentity.Mspid mathces the ID of
the MSP that performs the deserialization.

Change-Id: Ie76c500e93b9acf1ea9c69ccc8db46f1e830d204
Signed-off-by: Alessandro Sorniotti <ale.linux@sopit.net>