[FAB-1946] Rm ChainHeader from ConfigurationItem
https://jira.hyperledger.org/browse/FAB-1946

The current configuration transaction format expects a header per
configuration item.  This is more flexible, as it allows multiple
parties to each sign subsets of a configuration, then submit it
together.  However, this additional workflow seems unnecessary, and it
multiplies the number of required signatures.  It also necessitates the
computation of a digest in the channel creation request which is hard on
the SDKs.  (This digest computation will be removed in a later CR)

For now, the ChainHeader is moved to the ConfigurationEnvelope, which
means it is not included in the signature, this is insecure, but is the
path of least complexity and smallest diff going forward until we reach
the ultimate secure solution via FAB-1880.

Change-Id: Idaf16c2afe0e2ebdd068d7b312a883a41a4a2017
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>