[FAB-1958] Add MSP test template
https://jira.hyperledger.org/browse/FAB-1958

This is part of an ongoing change series which has the individual
components automatically generate a template during tests of the
configuration that is requried in order to intialize said component.

This one adds a template for MSP which will hopefully allow us to
shortly rip out all of the LoadFake...MSP calls.

Change-Id: I91f455f70f6286aa2a8eccb32221d3fcf9199415
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>