[FAB-1992] Move configtx signatures to envelope
https://jira.hyperledger.org/browse/FAB-1992

The current configtx scheme is a collection of signed configuration
items. This has the benefit of being very flexible, allowing a submitter
to collect signatures from one set of parties for one set of items, a
second set of signatures from a second set of parties for a second set
of items, and then glue them together, and submit a valid
reconfiguration.

However, this flexibility comes at the expense of requiring more
signatures, and the added complexity does not seem worth the added
flexibility. This CR moves the signatures to the envelope level, but
maintains the per item policy evaluation.

This is a step on the way towards:

https://jira.hyperledger.org/browse/FAB-1880

Which will improve the administrative process even further.

Change-Id: Iae35ed8ac5577603666de9d79a50d51509ad2d45
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>