[FAB-2101] Move orderer sharedconfig to common
https://jira.hyperledger.org/browse/FAB-2101

It has always been somewhat problematic that orderer config generation
is done in the orderer package, even though other pieces of the system
depend on it.

This CR moves the orderer shared config definitions to the common
package.  This is necessary to define a config schema going forward in
order to lock down the pending more flexible configuration format.

Change-Id: I1b1a4c78eedbc18a90d6838af0829950d6d88e72
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>