[FAB-2105] Add config schema protos
https://jira.hyperledger.org/browse/FAB-2105

Because the new configtx format is so much more flexible than the
original, we need to make sure that it is locked down to keep people
from going crazy with it. The restrictions can be relaxed as the use
cases are made, but initially the goal is to restrict the configuration
to a minimal set and loosen from there.

To do this, protos for a simple schema should be defined so that the
configtx manager can enforce the restrictive configuration scheme.

Change-Id: Iff0df7910aa216a03267beeb86b5b9f3f29d4b70
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>