Fixes errors in channel_test.sh
Following the instructions in /docs/gettingstarted.md creates a fabric deployment that continuously logs errors related to security settings. These can be corrected by creating the channel with:

1. CORE_PEER_GOSSIP_IGNORESECURITY=true
2. An anchor file with the certificate associated with peer0

As the IP address of peer0 isn't known ahead of time, we need to generate the anchor file dynamically. This commit updates the script and the tarball referenced in the documentation.

Change-Id: Id236c64f5c4ed98981ac33b16083765c2de23301
Signed-off-by: Sean Wellington <sean@d20tech.com>
Signed-off-by: Ramesh babu <rameshbabu.thoomu@gmail.com>