[FAB-2122] Generalize GOLANG install path
We have a defunct configuration key to specify the chaincode
install path, but it is not used.  We also specify a chaincode
specific binary name, which is not really necessary.  Therefore
we remove the defunct config item and implement a generic
binary name to clean this up.

Part of the fix for FAB-2122.

Change-Id: I84f76862535f819e4f023b7822ec89b441e07a1e
Signed-off-by: Greg Haskins <gregory.haskins@gmail.com>