[FAB-2157] Retry connection to bootstrap peers
In gossip, if a peer starts up and the bootstrap peers aren't available,
a reconnection attempt isn't being made.
This is risky especially in automatically-managed environments
like the cloud where machines can come up automatically and
there is a large number of them and no one analyzes their logs in real-time.

I didn't add a new test, but I re-arranged a test in the discovery layer
and in the gossip layer and made the bootstrap peers start after the rest
of the peers.

Change-Id: Ib49179cfd4d17e1ed9c6b33ee522769ec4efc082
Signed-off-by: Yacov Manevich <yacovm@il.ibm.com>