Replay attack protection
This change-set enforces the field TxID of a
proposal/transaction to be computed as the result
of an hash function computed on nonce and creator.
This is to be able to uniquely identity a proposal/transaction.

In addition, the endorsers verify that a TxID is not duplicated
by looking up the ledger as done by the committers.

Before merging this change-set, the client-sdk
needs to be modified to generate the TxID in the
new expected way.

This change-set comes in the context of
https://jira.hyperledger.org/browse/FAB-583

Change-Id: Icc5892976aeec1d4fdca736234355833d04bd4c8
Signed-off-by: Angelo De Caro <adc@zurich.ibm.com>