[FAB-2024] Add valid indicator to GetTransactionById
When clients retrieve a transaction from ledger, they need to know whether the
transaction was validated or invalidated by committing peer.

GetTransactionByID API needs to retrieve the transaction Envelope
from block storage, and return it to a client, and indicate whether the
transaction was validated or invalidated by committing peer. So that the
originally submitted transaction Envelope is not modified, a new
ProcessedTransaction wrapper is returned.

Also removed InvalidTransaction, since it is not used.

Change-Id: I7a4ec2b56b120f5df4ef18f0df415bab1a7dc9d1
Signed-off-by: denyeart <enyeart@us.ibm.com>