[FAB-2198] Adjust gossip membership layer
Continuation of:
	https://gerrit.hyperledger.org/r/#/c/5903/
	Introduce envelopes to gossip message

The gossip membership module has a storage of GossipMessages.
It needs to be refactored to hold SignedMessages instead, but
SignedMessages don't have any meaningful methods because they are
only envelopes to GossipMessages.

Therefore, this commit introduces an internal message type to
the gossip membership layer, that has both the SignedGossipMessage
and the GossipMessage as pointers.

Also- previously the cachedMembership field was used but now
can't be used because it's a raw protobuf structure.
Therefore, I refactored it to a 'membershipStore' which is just
a nice wrapper around a map from PKI-ID --> message.

Signed-off-by: Yacov Manevich <yacovm@il.ibm.com>
Change-Id: I2251c912b6a4610f6565f51c67bbb89d5d26b94f