fab-2177 add install command to lccc
With the new deploy model where the installation is on the file system,
the "on peer" install is not quite satisfactory for developers who have
to log on to a peer to install before they do anything.

This implements a new "install" Invoke command on the LCCC whose only
purpose is to put the code on the file system.

CLI usage
---------
peer chaincode install -n <name> -v <version> -p <path>

where the command will install to CORE_PEER_ADDRESS.

Change-Id: Ieadd7b9495ddf57ab7099cb788d8b99f4553cdf6
Signed-off-by: Srinivasan Muralidharan <muralisr@us.ibm.com>