[FAB-2201] Overlay writeset onto existing config
https://jira.hyperledger.org/browse/FAB-2201

The configtx code currently shortcuts by assuming that the writeset
contains exactly the entire config. In order to support partial updates,
this writeset should be overlayed on top of the existing config, then
transformed back into a new config.

This CR still depends on the writeset containing the entire config, but,
readies for the conversion to a partial writeset.

Change-Id: I35a71a02437865abe4178c47adfb7c5479b86a22
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>