[FAB-2202] Initialize configtx.Manager from Config
https://jira.hyperledger.org/browse/FAB-2202

As a transitional mechanism, the config currently depends on the
WriteSet containing the entire configuration. Instead, the config as
written in the config envelope needs to be utilized.

This CR makes the configtx.Manager utilize the config envelope at
instantiation time, but the orderer multichannel support still needs to
be updated to generate the config envelope based on a WriteSet for
instantiation.

Change-Id: I35c069cfd1ca523fe6c80f6acccfabb0e5ed8ad0
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>