FAB-2203 chaincode init should follow invoke FSM paths
https://jira.hyperledger.org/browse/FAB-2203

Chaincode init was tied to deploy and had a different state
model where invokes cannot happen while the chaincode was being
initialized. With the change that separates launch from inits and
invokes, init has to follow the same state as invokes. Currently
init is sending message to chaincode OOB which caused a race
with a subsequent invoke TX.

Changing init to chaincode to follow the Invoke tx paths.

Change-Id: Ic091782fd013a3a0e7781ba67449340fb90696e6
Signed-off-by: Srinivasan Muralidharan <muralisr@us.ibm.com>