fab-2220 getcc record not to check for CC existence
https://jira.hyperledger.org/browse/FAB-2220

LCCC exposes the following "GET" api.
   * getid - gets the name given name (basically "exists")
   * getccdata - gets the record entry for the chaincode given name
   * getdepspec - gets the deployment spec for the chaincode given name

Of these only the later needs the chaincode to be available on the peer.
Other APIs should not look for the existence of the chaincode.

This is needed for TXs to be committed on peer even if the chaincode
is not available on it

Change-Id: I34e336128880ccd3d5f82733a17b6e15fb14ec50
Signed-off-by: Srinivasan Muralidharan <muralisr@us.ibm.com>