Use buildenv image for linter/protos
At one point in time, "testenv" was used for basic build-time
processes such as linter, protos, and unit-tests.  Overtime, the
requirements for unit-testing grew and more dependencies were added.
This had the side-effect of requiring all prerequisites for unit-testing
be met for the other uses such as "make protos".  This led to a
chicken-egg scenario when certain dependencies required the protos
to be updated before they could be built, but the proto compilation
required them to be built before it could run.

This is clearly not correct.

We fix this by making the image contruction more fine-grained.  We split
the basic functionality of "testenv" out into "buildenv".  We then
allow the linter/protos process and the testenv image to consume buildenv.

Change-Id: I5db05bac649c2fefe73b19ec06e44efe3d2b42fa
Signed-off-by: Gregory Haskins <gregory.haskins@gmail.com>