[FAB-2232] Rename test MSP to DEFAULT
https://jira.hyperledger.org/browse/FAB-2232

The FAB-1880 change series picked the name "SAMPLE" when instantiating
the sample MSP config when an ID was added to the GetLocalMspFromDir
function.  This apparently as a side effect breaks other parts fo the
system, so this CR sets this ID back to "DEFAULT" as other pieces of the
system expect.

Change-Id: I5df97693561ab53d47f34d14d381e4f2fe4a92ad
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>