FAB-1764 CouchDB name to be consistent with chain ID
Motivation for this change:
Database creation fails on CouchDB if chainid does not meet
Couch DB database naming conventions

Chainid becomes database name in CouchDB.

CouchDB naming conventions:

Only lowercase characters (a-z), digits (0-9), and any of the characters
 _, $, (, ), +, -,and / are allowed. Must begin with a letter.

Therefore, we either need to adopt these same restrictions for
chainid naming, or have a deterministic mapping of chainid to a valid
(and unique) CouchDB database name.

The chainid also becomes a directory on the peer file system for
the block storage ledger.
Therefore the character '/' must also be avoided.

Change-Id: I0582e63b53e2afb5dee2009ede5f695775282b46
Signed-off-by: Chris Elder <chris.elder@us.ibm.com>