[FAB-2257] Fix local orderer startup
https://jira.hyperledger.org/browse/FAB-2257

In fixing the orderer docker startup, the orderer.yaml file was modified
to include the LocalMSPDir.  This was required because without the
variable in the yaml, the environment variable would not take effect.
However, this orderer.yaml variable was incorrectly set to ../msp when
it was being set as a default in code to ../msp/sampleconfig.  This CR
fixes that.

Change-Id: I775b70a0d14d5a9b13ed71b71ef196a70748690a
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>