[FAB-2261] Make Handler creation transactional
https://jira.hyperledger.org/browse/FAB-2261

At the organization level, handlers were being instaniated dynamically,
but, the handlers were only being instructed to begin transactions
statically.  This caused a crash when a dynamically created org handler
would be told to process a config item before begin the config
transaction.

This CR reworks the handler instantiation to be done at the beginning of
a transaction.

Change-Id: Ifbd2f9adf5809cb9fca3108573f105db4b0d790a
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>