Use a minimal container for GOLANG/CAR chaincode
This patch implements a "pass-through" build of chaincode
within a docker container as an alternative to using standard
"docker build" + Dockerfile mechanisms.  The plain docker
build is somewhat limiting due to the resulting image that is
a superset composition of the build-time and run-time environments.
This superset can be problematic on several fronts, such as a
bloated image size, and additional security exposure associated with
applications that are not needed, etc.

Before this patch, chaincode containers were approximately 1.4GB.
After this patch, they are about 175MB.

Java chaincode has not been ported to this process

Change-Id: Ib4e9881a0c1bfd6fb2f098e105a61d82c5293b3d
Signed-off-by: Greg Haskins <gregory.haskins@gmail.com>